getContrastTextColor reducet to single definition
